+++ demo/IPC/server/server.c
@@ -19,7 +19,7 @@ int EAPP_ENTRY main(){
       sum += nums[i];
     }
   }
-  ret_arg.req_vaddr = vaddr;
+  ret_arg.req_vaddr = (unsigned long)vaddr;
   //ret_arg.req_vaddr = 0;
   ret_arg.req_size = size;
   eapp_print("server read req_vaddr:0x%lx\n", sum);
@@ -32,7 +32,7 @@ int EAPP_ENTRY main(){
     nums[i] = 2;
     sum += nums[i];
   }
-  ret_arg.resp_vaddr = nums;
+  ret_arg.resp_vaddr = (unsigned long)nums;
   //ret_arg.resp_vaddr = 0;
   ret_arg.resp_size = size;
   eapp_print("server read resp_vaddr:0x%lx\n", sum);
