+++ musl/src/unistd/lseek.c
@@ -5,19 +5,13 @@
 #include "fsclient.h"
 off_t __lseek(int fd, off_t offset, int whence)
 {
-// #ifdef SYS__llseek
-// 	off_t result;
-// 	return syscall(SYS__llseek, fd, offset>>32, offset, &result, whence) ? -1 : result;
-// #else
-// 	return syscall(SYS_lseek, fd, offset, whence);
-// #endif
 	struct lseek_arg *lseek_arg = (struct lseek_arg*)(get_pre_mapped_page());
 	lseek_arg->fd = fd;
 	lseek_arg->offset = offset;
 	lseek_arg->whence = whence;
 	struct call_enclave_arg_t call_arg;
 	call_arg.req_arg = FSLSEEK;
-	call_arg.req_vaddr = lseek_arg;
+	call_arg.req_vaddr = (unsigned long)lseek_arg;
 	call_arg.req_size = PRE_MAP_SIZE;
 	int status = call_enclave(get_fs_handle(),&call_arg);
 	set_pre_map_page(call_arg.req_vaddr);
